+++ components/search/src/configuration_types.rs
@@ -60,6 +60,10 @@ pub(crate) struct JSONEngineUrl {
     /// URL. The client will fall back to the special locale code "default" when
     /// its locale is not present in the map.
     pub display_name_map: Option<HashMap<String, String>>,
+
+    /// Indicates the date until which the URL is considered new
+    /// (format: YYYY-MM-DD).
+    pub is_new_until: Option<String>,
 }
 
 /// Reflects `types::SearchEngineUrls`, but using `EngineUrl`.