+++ components/search/src/selector.rs
@@ -513,7 +513,7 @@ mod tests {
                                     experiment_config: None
                                 }],
                                 search_term_param_name: Some("q".to_string()),
-                                display_name: None,
+                                ..Default::default()
                             },
                             suggestions: Some(SearchEngineUrl {
                                 base: "https://example.com/suggestions".to_string(),
@@ -525,7 +525,7 @@ mod tests {
                                     experiment_config: None
                                 }],
                                 search_term_param_name: Some("suggest".to_string()),
-                                display_name: None,
+                                ..Default::default()
                             }),
                             trending: Some(SearchEngineUrl {
                                 base: "https://example.com/trending".to_string(),
@@ -538,8 +538,7 @@ mod tests {
                                         "trending-experiment-value".to_string()
                                     )
                                 }],
-                                search_term_param_name: None,
-                                display_name: None,
+                                ..Default::default()
                             }),
                             search_form: Some(SearchEngineUrl {
                                 base: "https://example.com/search-form".to_string(),
@@ -550,8 +549,7 @@ mod tests {
                                     experiment_config: None,
                                     enterprise_value: None,
                                 }],
-                                search_term_param_name: None,
-                                display_name: None,
+                                ..Default::default()
                             }),
                             visual_search: Some(SearchEngineUrl {
                                 base: "https://example.com/visual-search".to_string(),
@@ -563,7 +561,7 @@ mod tests {
                                     enterprise_value: None,
                                 }],
                                 search_term_param_name: Some("url".to_string()),
-                                display_name: None,
+                                ..Default::default()
                             }),
                         },
                         ..Default::default()
@@ -582,10 +580,8 @@ mod tests {
                         urls: SearchEngineUrls {
                             search: SearchEngineUrl {
                                 base: "https://example.com/2".to_string(),
-                                method: "GET".to_string(),
-                                params: Vec::new(),
                                 search_term_param_name: Some("search".to_string()),
-                                display_name: None,
+                                ..Default::default()
                             },
                             suggestions: None,
                             trending: None,
@@ -771,7 +767,7 @@ mod tests {
                                     experiment_config: None
                                 }],
                                 search_term_param_name: Some("q".to_string()),
-                                display_name: None,
+                                ..Default::default()
                             },
                             suggestions: Some(SearchEngineUrl {
                                 base: "https://example.com/suggestions".to_string(),
@@ -783,7 +779,7 @@ mod tests {
                                     experiment_config: None
                                 }],
                                 search_term_param_name: Some("suggest".to_string()),
-                                display_name: None,
+                                ..Default::default()
                             }),
                             trending: Some(SearchEngineUrl {
                                 base: "https://example.com/trending".to_string(),
@@ -794,8 +790,7 @@ mod tests {
                                     enterprise_value: None,
                                     experiment_config: Some("area-param".to_string())
                                 }],
-                                search_term_param_name: None,
-                                display_name: None,
+                                ..Default::default()
                             }),
                             search_form: Some(SearchEngineUrl {
                                 base: "https://example.com/search-form".to_string(),
@@ -806,8 +801,7 @@ mod tests {
                                     enterprise_value: None,
                                     experiment_config: None,
                                 }],
-                                search_term_param_name: None,
-                                display_name: None,
+                                ..Default::default()
                             }),
                             visual_search: Some(SearchEngineUrl {
                                 base: "https://example.com/visual-search".to_string(),
@@ -819,7 +813,7 @@ mod tests {
                                     experiment_config: None,
                                 }],
                                 search_term_param_name: Some("url".to_string()),
-                                display_name: None,
+                                ..Default::default()
                             }),
                         },
                         ..Default::default()
@@ -835,10 +829,8 @@ mod tests {
                         urls: SearchEngineUrls {
                             search: SearchEngineUrl {
                                 base: "https://example.com/2".to_string(),
-                                method: "GET".to_string(),
-                                params: Vec::new(),
                                 search_term_param_name: Some("search".to_string()),
-                                display_name: None,
+                                ..Default::default()
                             },
                             ..Default::default()
                         },
@@ -1023,7 +1015,7 @@ mod tests {
                                 experiment_config: None
                             }],
                             search_term_param_name: Some("q".to_string()),
-                            display_name: None,
+                            ..Default::default()
                         },
                         suggestions: Some(SearchEngineUrl {
                             base: "https://example.com/suggestions".to_string(),
@@ -1035,7 +1027,7 @@ mod tests {
                                 experiment_config: None
                             }],
                             search_term_param_name: Some("suggest".to_string()),
-                            display_name: None,
+                            ..Default::default()
                         }),
                         trending: Some(SearchEngineUrl {
                             base: "https://example.com/trending".to_string(),
@@ -1047,7 +1039,7 @@ mod tests {
                                 experiment_config: Some("area-param".to_string())
                             }],
                             search_term_param_name: None,
-                            display_name: None,
+                            ..Default::default()
                         }),
                         search_form: Some(SearchEngineUrl {
                             base: "https://example.com/search-form".to_string(),
@@ -1059,7 +1051,7 @@ mod tests {
                                 experiment_config: None,
                             }],
                             search_term_param_name: None,
-                            display_name: None,
+                            ..Default::default()
                         }),
                         visual_search: Some(SearchEngineUrl {
                             base: "https://example.com/visual-search".to_string(),
@@ -1071,7 +1063,7 @@ mod tests {
                                 experiment_config: None,
                             }],
                             search_term_param_name: Some("url".to_string()),
-                            display_name: None,
+                            ..Default::default()
                         }),
                     },
                     ..Default::default()
@@ -1112,7 +1104,7 @@ mod tests {
                                 experiment_config: None
                             }],
                             search_term_param_name: Some("q".to_string()),
-                            display_name: None,
+                            ..Default::default()
                         },
                         suggestions: Some(SearchEngineUrl {
                             base: "https://example.com/suggestions".to_string(),
@@ -1124,7 +1116,7 @@ mod tests {
                                 experiment_config: None
                             }],
                             search_term_param_name: Some("suggest".to_string()),
-                            display_name: None,
+                            ..Default::default()
                         }),
                         trending: Some(SearchEngineUrl {
                             base: "https://example.com/trending".to_string(),
@@ -1136,7 +1128,7 @@ mod tests {
                                 experiment_config: Some("area-param".to_string())
                             }],
                             search_term_param_name: None,
-                            display_name: None,
+                            ..Default::default()
                         }),
                         search_form: Some(SearchEngineUrl {
                             base: "https://example.com/search-form".to_string(),
@@ -1148,7 +1140,7 @@ mod tests {
                                 experiment_config: None,
                             }],
                             search_term_param_name: None,
-                            display_name: None,
+                            ..Default::default()
                         }),
                         visual_search: Some(SearchEngineUrl {
                             base: "https://example.com/visual-search".to_string(),
@@ -1160,7 +1152,7 @@ mod tests {
                                 experiment_config: None,
                             }],
                             search_term_param_name: Some("url".to_string()),
-                            display_name: None,
+                            ..Default::default()
                         }),
                     },
                     ..Default::default()
@@ -1301,10 +1293,8 @@ mod tests {
                         urls: SearchEngineUrls {
                             search: SearchEngineUrl {
                                 base: "https://example.com/1".to_string(),
-                                method: "GET".to_string(),
-                                params: Vec::new(),
                                 search_term_param_name: Some("q".to_string()),
-                                display_name: None,
+                                ..Default::default()
                             },
                             ..Default::default()
                         },
@@ -1339,10 +1329,8 @@ mod tests {
                         urls: SearchEngineUrls {
                             search: SearchEngineUrl {
                                 base: "https://example.com/1".to_string(),
-                                method: "GET".to_string(),
-                                params: Vec::new(),
                                 search_term_param_name: Some("q".to_string()),
-                                display_name: None,
+                                ..Default::default()
                             },
                             ..Default::default()
                         },
@@ -1356,10 +1344,8 @@ mod tests {
                         urls: SearchEngineUrls {
                             search: SearchEngineUrl {
                                 base: "https://example.com/2".to_string(),
-                                method: "GET".to_string(),
-                                params: Vec::new(),
                                 search_term_param_name: Some("search".to_string()),
-                                display_name: None,
+                                ..Default::default()
                             },
                             ..Default::default()
                         },
@@ -1395,10 +1381,8 @@ mod tests {
                         urls: SearchEngineUrls {
                             search: SearchEngineUrl {
                                 base: "https://example.com/1".to_string(),
-                                method: "GET".to_string(),
-                                params: Vec::new(),
                                 search_term_param_name: Some("q".to_string()),
-                                display_name: None,
+                                ..Default::default()
                             },
                             ..Default::default()
                         },
@@ -1412,10 +1396,8 @@ mod tests {
                         urls: SearchEngineUrls {
                             search: SearchEngineUrl {
                                 base: "https://example.com/3".to_string(),
-                                method: "GET".to_string(),
-                                params: Vec::new(),
                                 search_term_param_name: Some("trek".to_string()),
-                                display_name: None,
+                                ..Default::default()
                             },
                             ..Default::default()
                         },
@@ -1550,10 +1532,7 @@ mod tests {
             urls: SearchEngineUrls {
                 search: SearchEngineUrl {
                     base: "https://example.com".to_string(),
-                    method: "GET".to_string(),
-                    params: Vec::new(),
-                    search_term_param_name: None,
-                    display_name: None,
+                    ..Default::default()
                 },
                 ..Default::default()
             },
@@ -1567,10 +1546,7 @@ mod tests {
             urls: SearchEngineUrls {
                 search: SearchEngineUrl {
                     base: "https://example.com".to_string(),
-                    method: "GET".to_string(),
-                    params: Vec::new(),
-                    search_term_param_name: None,
-                    display_name: None,
+                    ..Default::default()
                 },
                 ..Default::default()
             },
@@ -1584,10 +1560,7 @@ mod tests {
             urls: SearchEngineUrls {
                 search: SearchEngineUrl {
                     base: "https://example.com".to_string(),
-                    method: "GET".to_string(),
-                    params: Vec::new(),
-                    search_term_param_name: None,
-                    display_name: None,
+                    ..Default::default()
                 },
                 ..Default::default()
             },
@@ -2462,15 +2435,13 @@ mod tests {
             urls: SearchEngineUrls {
                 search: SearchEngineUrl {
                     base: "https://example.com/search-overrides".to_string(),
-                    method: "GET".to_string(),
                     params: vec![SearchUrlParam {
                         name: "overrides-name".to_string(),
                         value: Some("overrides-value".to_string()),
                         enterprise_value: None,
                         experiment_config: None,
                     }],
-                    search_term_param_name: None,
-                    display_name: None,
+                    ..Default::default()
                 },
                 ..Default::default()
             },
@@ -2520,10 +2491,7 @@ mod tests {
             urls: SearchEngineUrls {
                 search: SearchEngineUrl {
                     base: "https://example.com".to_string(),
-                    method: "GET".to_string(),
-                    params: Vec::new(),
-                    search_term_param_name: None,
-                    display_name: None,
+                    ..Default::default()
                 },
                 ..Default::default()
             },
@@ -2537,10 +2505,7 @@ mod tests {
             urls: SearchEngineUrls {
                 search: SearchEngineUrl {
                     base: "https://example.com".to_string(),
-                    method: "GET".to_string(),
-                    params: Vec::new(),
-                    search_term_param_name: None,
-                    display_name: None,
+                    ..Default::default()
                 },
                 ..Default::default()
             },
@@ -2554,10 +2519,7 @@ mod tests {
             urls: SearchEngineUrls {
                 search: SearchEngineUrl {
                     base: "https://example.com".to_string(),
-                    method: "GET".to_string(),
-                    params: Vec::new(),
-                    search_term_param_name: None,
-                    display_name: None,
+                    ..Default::default()
                 },
                 ..Default::default()
             },
@@ -2637,10 +2599,7 @@ mod tests {
             urls: SearchEngineUrls {
                 search: SearchEngineUrl {
                     base: "https://example.com".to_string(),
-                    method: "GET".to_string(),
-                    params: Vec::new(),
-                    search_term_param_name: None,
-                    display_name: None,
+                    ..Default::default()
                 },
                 ..Default::default()
             },
@@ -2654,10 +2613,7 @@ mod tests {
             urls: SearchEngineUrls {
                 search: SearchEngineUrl {
                     base: "https://example.com".to_string(),
-                    method: "GET".to_string(),
-                    params: Vec::new(),
-                    search_term_param_name: None,
-                    display_name: None,
+                    ..Default::default()
                 },
                 ..Default::default()
             },
@@ -2806,15 +2762,13 @@ mod tests {
             urls: SearchEngineUrls {
                 search: SearchEngineUrl {
                     base: "https://example.com/search-overrides".to_string(),
-                    method: "GET".to_string(),
                     params: vec![SearchUrlParam {
                         name: "overrides-name".to_string(),
                         value: Some("overrides-value".to_string()),
                         enterprise_value: None,
                         experiment_config: None,
                     }],
-                    search_term_param_name: None,
-                    display_name: None,
+                    ..Default::default()
                 },
                 ..Default::default()
             },
@@ -2831,10 +2785,7 @@ mod tests {
             urls: SearchEngineUrls {
                 search: SearchEngineUrl {
                     base: "https://example.com/search-distro".to_string(),
-                    method: "GET".to_string(),
-                    params: Vec::new(),
-                    search_term_param_name: None,
-                    display_name: None,
+                    ..Default::default()
                 },
                 ..Default::default()
             },
@@ -2953,10 +2904,7 @@ mod tests {
             urls: SearchEngineUrls {
                 search: SearchEngineUrl {
                     base: "https://example.com".to_string(),
-                    method: "GET".to_string(),
-                    params: Vec::new(),
-                    search_term_param_name: None,
-                    display_name: None,
+                    ..Default::default()
                 },
                 ..Default::default()
             },
@@ -2970,10 +2918,7 @@ mod tests {
             urls: SearchEngineUrls {
                 search: SearchEngineUrl {
                     base: "https://example.com".to_string(),
-                    method: "GET".to_string(),
-                    params: Vec::new(),
-                    search_term_param_name: None,
-                    display_name: None,
+                    ..Default::default()
                 },
                 ..Default::default()
             },